added error messages and notifications

+++ src/main/kotlin/mbraun/student_application/service/StudentService.kt
@@ -1,5 +1,7 @@
 package mbraun.student_application.service
 
+import mbraun.student_application.exception.BadRequestException
+import mbraun.student_application.exception.StudentNotFoundException
 import mbraun.student_application.model.Student
 import mbraun.student_application.repository.StudentRepository
 import org.springframework.stereotype.Service
@@ -12,10 +14,18 @@ class StudentService(val studentRepository: StudentRepository) {
     }
 
     fun addStudent(student: Student) {
+        val existsEmail = studentRepository.selectExistsEmail(student.email)
+
+        if (existsEmail) {
+            throw BadRequestException("The Email ${student.email} is already taken")
+        }
         studentRepository.save(student)
     }
 
     fun deleteStudent(studentId: Long) {
+        if (!studentRepository.existsById(studentId)) {
+            throw StudentNotFoundException("The Student with id $studentId does not exist")
+        }
         studentRepository.deleteById(studentId)
     }
 }
